refactor back arquitectura
se cambia a arquitectura modular/features ya que implementar hexagonal no llevaría mas tiempo para implementar el MVP y agregar nuevas features
no se borra nada, solo se reubican los archivos, esto se debe a que se hizo el cambio a tiempo, antes de implementar patrones hexagonales
