+++ backend/src/main/java/com/team42/churninsight/common/dto/ErrorResponse.java
@@ -1,0 +1,11 @@
+package com.team42.churninsight.common.dto;
+
+import java.time.LocalDateTime;
+
+public record ErrorResponse(
+        int status,
+        String error,
+        String message,
+        LocalDateTime timeStamp
+) {
+}
