+++ backend/src/main/java/com/team42/churninsight/common/exception/ModelUnavailableException.java
@@ -1,0 +1,7 @@
+package com.team42.churninsight.common.exception;
+
+public class ModelUnavailableException extends RuntimeException {
+    public ModelUnavailableException(String message) {
+        super(message);
+    }
+}
