+++ backend/src/main/java/com/team42/churninsight/common/exception/PersistenceException.java
@@ -1,0 +1,7 @@
+package com.team42.churninsight.common.exception;
+
+public class PersistenceException extends RuntimeException {
+    public PersistenceException(String message) {
+        super(message);
+    }
+}
