+++ backend/src/main/java/com/team42/churninsight/common/exception/PredictionFailedException.java
@@ -1,0 +1,7 @@
+package com.team42.churninsight.common.exception;
+
+public class PredictionFailedException extends RuntimeException {
+    public PredictionFailedException(String message) {
+        super(message);
+    }
+}
