--- backend/src/main/java/com/team42/churninsight/domain/enums/Churn.java
@@ -1,6 +1,0 @@
-package com.team42.churninsight.domain.enums;
-
-public enum Churn {
-    CHURN,
-    NO_CHURN
-}
