--- backend/src/main/java/com/team42/churninsight/domain/enums/Gender.java
@@ -1,9 +1,0 @@
-package com.team42.churninsight.domain.enums;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum Gender {
-    @JsonProperty("Male") MALE,
-    @JsonProperty("Female") FEMALE,
-    @JsonProperty("Other") OTHER
-}
