--- backend/src/main/java/com/team42/churninsight/domain/enums/IncomeBracket.java
@@ -1,9 +1,0 @@
-package com.team42.churninsight.domain.enums;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum IncomeBracket {
-    @JsonProperty("Low") LOW,
-    @JsonProperty("Medium") MEDIUM,
-    @JsonProperty("High") HIGH
-}
