--- backend/src/main/java/com/team42/churninsight/domain/enums/MaritalStatus.java
@@ -1,9 +1,0 @@
-package com.team42.churninsight.domain.enums;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum MaritalStatus {
-    @JsonProperty("Divorced") DIVORCED,
-    @JsonProperty("Single") SINGLE,
-    @JsonProperty("Married") MARRIED
-}
