--- backend/src/main/java/com/team42/churninsight/infrastructure/exception/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.team42.churninsight.infrastructure.exception;
-
-import java.time.LocalDateTime;
-
-public record ErrorResponse(
-        int status,
-        String error,
-        String message,
-        LocalDateTime timeStamp
-) {
-}
