--- backend/src/main/java/com/team42/churninsight/infrastructure/exception/ModelUnavailableException.java
@@ -1,7 +1,0 @@
-package com.team42.churninsight.infrastructure.exception;
-
-public class ModelUnavailableException extends RuntimeException {
-    public ModelUnavailableException(String message) {
-        super(message);
-    }
-}
