--- backend/src/main/java/com/team42/churninsight/infrastructure/exception/PersistenceException.java
@@ -1,7 +1,0 @@
-package com.team42.churninsight.infrastructure.exception;
-
-public class PersistenceException extends RuntimeException {
-    public PersistenceException(String message) {
-        super(message);
-    }
-}
