--- backend/src/main/java/com/team42/churninsight/infrastructure/exception/PredictionFailedException.java
@@ -1,7 +1,0 @@
-package com.team42.churninsight.infrastructure.exception;
-
-public class PredictionFailedException extends RuntimeException {
-    public PredictionFailedException(String message) {
-        super(message);
-    }
-}
