--- backend/src/main/java/com/team42/churninsight/ports/in/README.md
@@ -1,7 +1,0 @@
-Por qué NO conviene ports/in ahora
-
-Añade una capa extra (interfaz + implementación) por cada caso de uso
-
-El beneficio real es pequeño si solo hay REST
-
-Puede confundir por ahora
