+++ backend/src/main/java/com/team42/churninsight/prediction/Prediction.java
@@ -1,0 +1,41 @@
+package com.team42.churninsight.prediction;
+
+
+import com.team42.churninsight.prediction.enums.Churn;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import java.time.LocalDateTime;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Prediction {
+    private Long id;
+    private String customerId;
+    private String transactionId;
+    private Double probabilityChurn;
+    private Churn churn;
+    private LocalDateTime createdAt;
+
+    public void calculateChurnStatus() {
+        if (this.probabilityChurn != null && this.probabilityChurn >= 0.5){
+            this.churn = Churn.CHURN;
+        }else {
+            this.churn = Churn.NO_CHURN;
+        }
+    }
+
+    public static Prediction create (String customerId, String transactionId, Double probabilityChurn){
+        Prediction newPrediction = new Prediction();
+        newPrediction.setCustomerId(customerId);
+        newPrediction.setTransactionId(transactionId);
+        newPrediction.setProbabilityChurn(probabilityChurn);
+        newPrediction.calculateChurnStatus();
+        newPrediction.setCreatedAt(LocalDateTime.now());
+        return newPrediction;
+    }
+
+}
+
