+++ backend/src/main/java/com/team42/churninsight/prediction/PredictionEntity.java
@@ -1,0 +1,29 @@
+package com.team42.churninsight.prediction;
+
+import com.team42.churninsight.prediction.enums.Churn;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "predictions")
+public class PredictionEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(name = "customer_id")
+    private String customerId;
+    @Column(name = "transaction_id")
+    private String transactionId;
+    private Double probability;
+    @Enumerated(EnumType.STRING)
+    private Churn churn;
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+}
