+++ backend/src/main/java/com/team42/churninsight/prediction/client/.gitkeep
@@ -1,0 +1,1 @@
+El cliente que se comunica con FastAPI
