+++ backend/src/main/java/com/team42/churninsight/prediction/enums/Churn.java
@@ -1,0 +1,6 @@
+package com.team42.churninsight.prediction.enums;
+
+public enum Churn {
+    CHURN,
+    NO_CHURN
+}
