+++ backend/src/main/java/com/team42/churninsight/prediction/enums/EducationLevel.java
@@ -1,0 +1,14 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum EducationLevel {
+    @JsonProperty("Master's")
+    MASTER_S,
+    @JsonProperty("Bachelor's")
+    BACHELOR_S,
+    @JsonProperty("High School")
+    HIGH_SCHOOL,
+    @JsonProperty("PhD")
+    PHD
+}
