+++ backend/src/main/java/com/team42/churninsight/prediction/enums/Gender.java
@@ -1,0 +1,9 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum Gender {
+    @JsonProperty("Male") MALE,
+    @JsonProperty("Female") FEMALE,
+    @JsonProperty("Other") OTHER
+}
