+++ backend/src/main/java/com/team42/churninsight/prediction/enums/IncomeBracket.java
@@ -1,0 +1,9 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum IncomeBracket {
+    @JsonProperty("Low") LOW,
+    @JsonProperty("Medium") MEDIUM,
+    @JsonProperty("High") HIGH
+}
