+++ backend/src/main/java/com/team42/churninsight/prediction/enums/MaritalStatus.java
@@ -1,0 +1,9 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum MaritalStatus {
+    @JsonProperty("Divorced") DIVORCED,
+    @JsonProperty("Single") SINGLE,
+    @JsonProperty("Married") MARRIED
+}
