+++ backend/src/main/java/com/team42/churninsight/prediction/enums/Occupation.java
@@ -1,0 +1,11 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum Occupation {
+    @JsonProperty("Retired") RETIRED,
+    @JsonProperty("Unemployed") UNEMPLOYED,
+    @JsonProperty("Student") STUDENT,
+    @JsonProperty("Employed") EMPLOYED,
+    @JsonProperty("Self-Employed") SELF_EMPLOYED
+}
