+++ backend/src/main/java/com/team42/churninsight/prediction/enums/ProductCategory.java
@@ -1,0 +1,11 @@
+package com.team42.churninsight.prediction.enums;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public enum ProductCategory {
+    @JsonProperty("Electronics") ELECTRONICS,
+    @JsonProperty("Toys") TOYS,
+    @JsonProperty("Groceries") GROCERIES,
+    @JsonProperty("Home Goods") HOME_GOODS,
+    @JsonProperty("Clothing") CLOTHING
+}
